complete service classes

+++ src/main/java/com/danglinh/project_bookstore/domain/entity/Feedback.java
@@ -12,7 +12,7 @@ public class Feedback {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "feedback_id")
-    private long feedbackId;
+    private int feedbackId;
 
     @Column(name = "feedback", columnDefinition = "text")
     private String feedback;
